+++ ReleaseNotes.md
@@ -1,5 +1,9 @@
 # FindFiles Release Notes
 
+**0.00.07**
+- styling update
+- fixed vflow reset
+
 **0.00.06**
 - show number of found files in the status string
 - added hit table to detail pane